peruocc: Query and Standardize Biodiversity Occurrences in Peru

Facilitates the retrieval, spatial validation, and integration of flora and fauna occurrence records across administrative units (districts and provinces) in Peru. Retrieves official boundary geometries via 'geoperu', queries and consolidates observations from the Global Biodiversity Information Facility (GBIF, https://www.gbif.org/) and 'iNaturalist' (https://www.inaturalist.org/), and standardizes attributes into a unified Darwin Core aligned structure. Designed for biodiversity assessments and spatial workflows within user-defined areas of interest.

See also

Author

Maintainer: Paul E. Santos Andrade paulefrens@gmail.com (ORCID) [copyright holder]

Authors:

Paul E. Santos Andrade paulefrens@gmail.com (ORCID) [copyright holder]